+++ Changelog.md
@@ -3,7 +3,7 @@
 ## 4.1.0 (2013-xx-xx)
 
  * IrcBase osztályban létrehozva a LoadProcessMethods() és UnloadProcessMethods(). Így egységesebb lesz a program több része is. Javítva az addonok leválasztása.
- * Runtime osztályhoz hozzáadva a MemorySize és MemorySizeToMB változó
+ * Runtime osztályhoz hozzáadva a MemorySize és MemorySizeInMB változó
  * Reload részhez hozzáadva az irc rész.
  * Névnapok kibövítésre kerültek az alábbi nyelveken: czCZ, deDE, fiFI, frFR, huHU, lvLV, plPL, seSE
  * CompilerAddon yaml konfig olvasása kijavítva.